Example Svk new attribute Svedala_RA.xml
This is an example of the added attribute normalEnabled that SvK wants to add for controllability on pins (inputs) and actions (controlling if they are possible to activate/be executed in the SIPS). See also SSI file for the attribute Enabled.

The altered pins and actions have been chosen randomly, they could be on any pins or actions. But if a pin or action does not have the attribute normaEnabled that means we do not have the described controllability.